requirements.txt
Django==1.11.4
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-geoposition==0.3.0
django-rest-framework==0.1.0
djangorestframework==3.6.4
easy-thumbnails==2.4.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
